Page d'accueil devient page de login, et connection d'utilisateurs possible
