+++ Views/Cagnotte/Index.cshtml
@@ -7,3 +7,5 @@
 <h1 class="header-cagnotte">Les Cagnottes Solidaires</h1>
 <h6 class ="header-cagnotte"><a href="/cagnotte/AnciennesCagnottes">(Anciennes cagnottes)</a></h6>
 @Html.Partial("ListeCagnottes", Model)
+
+Hello @User.Identity.Name
